Bruker ikke sammensatt navn da denne har max-lengde på 25 fra TPS
Flytter getNavn til person-objektet og bruker denne globalt
Oppdaterer tester
