+++ src/components/person/NavnOgAlder.tsx
@@ -1,31 +1,12 @@
 import * as React from 'react';
-import { erDød, Familierelasjon, Navn } from '../../models/person/person';
+import { erDød, Familierelasjon, getNavn } from '../../models/person/person';
 
 interface Props {
     relasjon: Familierelasjon;
 }
 
-function getNavn({fornavn, mellomnavn, etternavn, sammensatt}: Navn) {
-    if (!fornavn && !etternavn) {
-        return sammensatt || 'Ukjent navn';
-    }
-
-    let navn = [];
-    if (fornavn) {
-        navn.push(fornavn);
-    }
-    if (mellomnavn) {
-        navn.push(mellomnavn);
-    }
-    if (etternavn) {
-        navn.push(etternavn);
-    }
-
-    return navn.join(' ');
-}
-
 function getAlder(relasjon: Familierelasjon) {
-    if (relasjon.tilPerson.alder  === undefined || relasjon.tilPerson.alder === null) {
+    if (relasjon.tilPerson.alder === undefined || relasjon.tilPerson.alder === null) {
         return null;
     } else if (relasjon.tilPerson.alder > 0) {
         return relasjon.tilPerson.alder;
